AXI-DMA-specific:  Disable IRQDelay interrupt when opening.
Per Xilinx PG 021:

Interrupt Delay Time Out. This value is used for setting the interrupt timeout
value. The interrupt timeout mechanism causes the DMA engine to generate an
interrupt after the delay time period has expired. Timer begins counting at the
end of a packet and resets with receipt of a new packet or a timeout event
occurs.

Without disabling this, receiving a packet may fail if that packet is not
immediately available at the moment the AXI DMA is enabled.